requirements.txt
streamlit
moviepy
numpy
chromadb
torch
transformers
pillow
pandas
decord
pandas
ftfy
regex
tqdm
opencv-python
functional
timm
open_clip_torch
